# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WaytoAlgo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List user investments
      responses:
        "200":
          description: List of investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create a new investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvestmentBody"
      responses:
        "201":
          description: Investment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /investments/{id}:
    get:
      operationId: getInvestment
      tags: [investments]
      summary: Get investment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
...
components:
  schemas:
...
    Investment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        planTier:
          type: string
          enum: [tier1, tier2, tier3]
        dailyRate:
          type: number
        durationDays:
          type: integer
        earnedSoFar:
          type: number
        remainingDays:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
        hyperCoinAmount:
          type: number
        usdtAmount:
          type: number
        createdAt:
          type: string
      required: [id, userId, amount, planTier, dailyRate, durationDays, earnedSoFar, remainingDays, startDate, endDate, status, hyperCoinAmount, usdtAmount, createdAt]
    CreateInvestmentBody:
      type: object
      properties:
        amount:
          type: number
        hyperCoinAmount:
          type: number
        usdtAmount:
          type: number
      required: [amount, hyperCoinAmount, usdtAmount]
...